SFB-85: split up the creating of the field data for one fieldSubject
